# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SingAI API local
  version: 2.0.0
  description: |
    Contrato verificable de los tres servicios Express incluidos en el repositorio.
    La interfaz desplegada usa Firebase directamente; esas operaciones no son
    endpoints HTTP propios y por ello no forman parte de este contrato.
servers:
  - url: http://localhost:4100
    description: Servicio núcleo y API pública local
security: []
...
paths:
...
  /internal/profiles:
    post:
      tags: [Interna - perfil]
      summary: Crea el perfil asociado a una cuenta local
      servers: [{ url: http://localhost:4101 }]
      security: [{ serviceKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, displayName]
              properties:
                userId: { type: integer }
                displayName: { type: string, maxLength: 60 }
                learningGoal: { type: string, maxLength: 40 }
      responses:
        '201': { description: Perfil creado o ya existente }
        '403': { $ref: '#/components/responses/Error' }
  /internal/profile:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      tags: [Interna - perfil]
      summary: Obtiene un perfil por usuario
      servers: [{ url: http://localhost:4101 }]
      security: [{ serviceKey: [] }]
      responses:
        '200': { description: Perfil }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
    patch:
      tags: [Interna - perfil]
      summary: Actualiza un perfil por usuario
      servers: [{ url: http://localhost:4101 }]
      security: [{ serviceKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfilePatch'
      responses:
        '200': { description: Perfil actualizado }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /internal/progress:
    get:
      tags: [Interna - perfil]
      summary: Obtiene progreso por usuario
      servers: [{ url: http://localhost:4101 }]
      security: [{ serviceKey: [] }]
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200': { description: Perfil y lecciones }
        '403': { $ref: '#/components/responses/Error' }
  /internal/progress/complete:
    post:
      tags: [Interna - perfil]
      summary: Registra una lección por usuario
      servers: [{ url: http://localhost:4101 }]
      security: [{ serviceKey: [] }]
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteLessonRequest'
      responses:
        '200': { description: Progreso actualizado }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceKey:
      type: apiKey
      in: header
      name: x-service-key
  parameters:
    UserId:
      name: x-user-id
      in: header
      required: true
      schema: { type: integer }
...
  responses:
    Error:
      description: Error de validación, autorización o dependencia
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    ProfilePatch:
      type: object
      properties:
        displayName: { type: string, maxLength: 60 }
        avatar: { type: string }
        learningGoal: { type: string, maxLength: 40 }
        dailyGoal: { type: integer, minimum: 1, maximum: 10 }
        weeklyGoal: { type: integer, minimum: 1, maximum: 7 }
        experienceLevel:
          type: string
          enum: [principiante, intermedio, avanzado]
        preferredCategory:
          type: string
          enum: [abecedario, conversaciones, emociones, dias, acciones]
        bio: { type: string, maxLength: 180 }
        autoplayVideos: { type: boolean }
    CompleteLessonRequest:
      type: object
      required: [lessonId, score]
      properties:
        lessonId: { type: string, maxLength: 80 }
        score: { type: number, minimum: 0, maximum: 100 }